requirements.txt
pytest==9.0.2
python-dotenv>=1.0.1,<2.0.0
python-jose>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
